requirements.txt
aiogram==2.19
telethon
termcolor
SpeechRecognition
requests
GitPython